+++ alltheutils/decoy/__init__.py
@@ -1,0 +1,1 @@
+"""Bukas na lang kita mamahalin."""